requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    "make update-requirements"
#
bolton-clack==0.3.4
    # via -r requirements.in
bolton-eris==0.2.3
    # via
    #   -r requirements.in
    #   bolton-clack
    #   bolton-proctor
    #   magodo
    #   potoroo
bolton-ion==0.1.0
    # via bolton-eris
bolton-logrus==0.1.0
    # via
    #   -r requirements.in
    #   bolton-clack
bolton-metaman==0.1.2
    # via
    #   -r requirements.in
    #   bolton-eris
    #   bolton-logrus
    #   magodo
bolton-proctor==0.2.2
    # via vimala
bolton-typist==0.2.0
    # via
    #   -r requirements.in
    #   bolton-clack
    #   bolton-eris
    #   magodo
    #   vimala
commonmark==0.9.1
    # via rich
greenlet==1.1.2
    # via sqlalchemy
magodo==1.1.0
    # via -r requirements.in
potoroo==0.4.1
    # via -r requirements.in
pydantic==1.9.1
    # via
    #   bolton-clack
    #   sqlmodel
pygments==2.12.0
    # via rich
python-dateutil==2.8.2
    # via -r requirements.in
pyyaml==6.0
    # via bolton-clack
rich==12.4.4
    # via
    #   -r requirements.in
    #   textual
six==1.16.0
    # via python-dateutil
sqlalchemy==1.4.35
    # via
    #   -r requirements.in
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a23
    # via sqlmodel
sqlmodel==0.0.6
    # via -r requirements.in
structlog==21.5.0
    # via bolton-logrus
textual==0.1.18
    # via
    #   -r requirements.in
    #   textual-inputs
textual-inputs==0.2.5
    # via -r requirements.in
typing-extensions==4.2.0
    # via
    #   pydantic
    #   rich
    #   sqlalchemy2-stubs
    #   textual
vimala==0.1.2
    # via -r requirements.in
